chore: generalize keyboard shortcut and improve wording in example
