+++ polaris-react/src/components/Tooltip/components/TooltipOverlay/tests/TooltipOverlay.test.tsx
@@ -2,7 +2,6 @@ import React from 'react';
 import {mountWithApp} from 'tests/utilities';
 
 import {TooltipOverlay} from '../TooltipOverlay';
-import {Text} from '../../../../Text';
 
 describe('<TooltipOverlay />', () => {
   const defaultProps = {
@@ -51,40 +50,6 @@ describe('<TooltipOverlay />', () => {
 
     expect(tooltipOverlay).toContainReactComponent('div', {
       style: expect.objectContaining({zIndex: 100}),
-    });
-  });
-
-  it('does not add a keyboard shortcut element by default', () => {
-    const activator = document.createElement('button');
-    const tooltipOverlay = mountWithApp(
-      <TooltipOverlay {...defaultProps} activator={activator}>
-        Content
-      </TooltipOverlay>,
-    );
-
-    expect(tooltipOverlay).not.toContainReactComponent(Text, {
-      as: 'span',
-      variant: 'bodyMd',
-      color: 'subdued',
-    });
-  });
-
-  it('adds a keyboard shortcut if present', () => {
-    const activator = document.createElement('button');
-    const tooltipOverlay = mountWithApp(
-      <TooltipOverlay
-        {...defaultProps}
-        activator={activator}
-        keyboardShortcut="#B"
-      >
-        Content
-      </TooltipOverlay>,
-    );
-
-    expect(tooltipOverlay).toContainReactComponent(Text, {
-      as: 'span',
-      variant: 'bodyMd',
-      color: 'subdued',
     });
   });
 });
